+++ Data/FlagManager.hs
@@ -39,18 +39,39 @@ decodeFlag aFlag = decodeFlagSub (bitLen# -# 1#)
   where
     !(I# bitLen#) = bitLen
     decodeFlagSub (-1#) = []
-    decodeFlagSub idx# =
+    decodeFlagSub  idx# =
       if testBit aFlag (I# idx#)
         then toEnum (I# idx#) : decodeFlagSub (idx# -# 1#)
         else                    decodeFlagSub (idx# -# 1#)
 
 showFlag :: Flag -> String
-showFlag = showFlagSub (bitLen-1)
-showFlagSub aFlag (-1) = ""
-showFlagSub aFlag num =
-  if testBit aFlag num
-    then '1' : showFlagSub aFlag (num-1)
-    else '0' : showFlagSub aFlag (num-1)
+showFlag aFlag = showFlagSub (bitLen#-#1#)
+  where
+    !(I# bitLen#) = bitLen
+    showFlagSub (-1#) = ""
+    showFlagSub  idx# =
+      if testBit aFlag (I# idx#)
+        then '1' : showFlagSub (idx#-#1#)
+        else '0' : showFlagSub (idx#-#1#)
+
+showFlagFit :: (Bounded a, Enum a) => a -> Flag -> String
+showFlagFit a aFlag = showFlagSub bitLen#
+  where
+    !(I# bitLen#) = fromEnum (maxBound `asTypeOf` a)
+    showFlagSub (-1#) = ""
+    showFlagSub  idx# =
+      if testBit aFlag (I# idx#)
+        then '1' : showFlagSub (idx#-#1#)
+        else '0' : showFlagSub (idx#-#1#)
+
+showFlagBy :: Int -> Flag -> String
+showFlagBy (I# bitLen#) aFlag = showFlagSub (bitLen#-#1#)
+  where
+    showFlagSub (-1#) = ""
+    showFlagSub  idx# =
+      if testBit aFlag (I# idx#)
+        then '1' : showFlagSub (idx#-#1#)
+        else '0' : showFlagSub (idx#-#1#)
 
 -- eq f1 f2 = xor f1 f2 == zeroBits
 include f1 f2 = ((complement f1) .&. (f1 .|. f2)) == zeroBits