+++ ChangeLog.md
@@ -1,9 +1,13 @@
 # Revision history for FlagManager
 
+## 0.1.1.0  -- 2017/01/21
+
+* Implement `showFlagFit` and `showFlagBy`.
+
 ## 0.1.0.0  -- 2017/01/20
 
 * Functions' name are changed.
-* Implement `decodeFlag`.
+* Implement `decodeFlag` and `showFlag`.
 
 ## 0.0.0.0  -- 2017/01/18